+++ modules/edit/beule/bang5.c
@@ -99,8 +99,8 @@ MOD_ABILITY  module_ability =
 };
 
 static SERVICE_FUNCTIONS *service;
-int (*busybox)(int lft);
-static void (*reset_busybox)(int lft, const char *string);
+short (*busybox)(short lft);
+static void (*reset_busybox)(short lft, const char *string);
 
 typedef unsigned char uchar;
 typedef unsigned int uint;
