+++ modules/import/iim.c
@@ -73,7 +73,7 @@ MOD_INFO module_info = {"Inshape-Format",
 /*		1, 8 und 24 Bit, unkomprimiert				*/
 /* -------------------------------------------------*/
 /* -------------------------------------------------*/
-int imp_module_main(GARGAMEL *smurf_struct)
+short imp_module_main(GARGAMEL *smurf_struct)
 {
 	char *buffer, *ziel, *pal,
 		 BitsPerPixel, DatenOffset;
