+++ modules/import/scodl.c
@@ -81,7 +81,7 @@ MOD_INFO module_info=
 
 
 
-int imp_module_main(GARGAMEL *smurf_struct)
+short imp_module_main(GARGAMEL *smurf_struct)
 {
 SMURF_PIC *picture;
 char *cpic, *bitmap_data, *decoded_pic, *decpic_copy, PicByte;
